app/src/main/java/hexlet/code/App.java
package hexlet.code;
import hexlet.code.controller.RootController;
import hexlet.code.util.NamedRoutes;
import io.javalin.Javalin;
import io.javalin.rendering.template.JavalinJte;
import lombok.extern.slf4j.Slf4j;
@Slf4j
public final class App {
    public static void main(String[] args) {
        Javalin app = getApp();
        app.start(7070);
    }
    public static Javalin getApp() {
        var app = Javalin.create(config -> {
            config.bundledPlugins.enableDevLogging();
            config.fileRenderer(new JavalinJte());
        });
        app.get(NamedRoutes.rootPath(), RootController::index);
        return app;
    }
}
